pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "autoclaude"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
autoclaude = "orchestrator.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["orchestrator*"]
